+++ packages/components/src/text-control/index.tsx
@@ -2,7 +2,6 @@
  * External dependencies
  */
 import type { ChangeEvent, ForwardedRef } from 'react';
-import clsx from 'clsx';
 
 /**
  * WordPress dependencies
@@ -16,7 +15,6 @@ import { forwardRef } from '@wordpress/element';
 import BaseControl from '../base-control';
 import type { WordPressComponentProps } from '../context';
 import type { TextControlProps } from './types';
-import { maybeWarnDeprecated36pxSize } from '../utils/deprecated-36px-size';
 
 function UnforwardedTextControl(
 	props: WordPressComponentProps< TextControlProps, 'input', false >,
@@ -25,7 +23,7 @@ function UnforwardedTextControl(
 	const {
 		// Prevent passing this to `input`.
 		__nextHasNoMarginBottom: _,
-		__next40pxDefaultSize = false,
+		__next40pxDefaultSize: __,
 		label,
 		hideLabelFromVision,
 		value,
@@ -40,12 +38,6 @@ function UnforwardedTextControl(
 	const onChangeValue = ( event: ChangeEvent< HTMLInputElement > ) =>
 		onChange( event.target.value );
 
-	maybeWarnDeprecated36pxSize( {
-		componentName: 'TextControl',
-		size: undefined,
-		__next40pxDefaultSize,
-	} );
-
 	return (
 		<BaseControl
 			label={ label }
@@ -55,9 +47,7 @@ function UnforwardedTextControl(
 			className={ className }
 		>
 			<input
-				className={ clsx( 'components-text-control__input', {
-					'is-next-40px-default-size': __next40pxDefaultSize,
-				} ) }
+				className="components-text-control__input"
 				type={ type }
 				id={ id }
 				value={ value }
@@ -82,7 +72,6 @@ function UnforwardedTextControl(
  *
  *   return (
  *     <TextControl
- *       __next40pxDefaultSize
  *       label="Additional CSS Class"
  *       value={ className }
  *       onChange={ ( value ) => setClassName( value ) }